+++ src/app/eth-memento/ethMementoPlugin.ts
@@ -8,7 +8,6 @@ import { blockTxsModule } from "app/shared/module/block/blockTxs/blockTxsModule"
 import { blockAdvancedModule } from "app/shared/module/block/blockAdvanced/blockAdvancedModule";
 import { blockLogsBloomModule } from "app/shared/module/block/blockLogsBloom/blockLogsBloomModule";
 import { BlockListAdapter } from "app/shared/adapter/block/BlockListAdapter";
-import { NullEthPriceAdapter } from "app/eth-memento/adapter/NullEthPriceAdapter";
 import { txParentBlockContext } from "app/shared/context/txParentBlockContext";
 import { txBasicModule } from "app/shared/module/tx/txBasic/txBasicModule";
 import { TxDetailsAdapter } from "app/eth-memento/adapter/tx/TxDetailsAdapter";
@@ -48,7 +47,6 @@ const ethMementoPlugin: IPlugin = {
             new LatestBlockNumberAdapter(dataSource.stores.blockStateStore));
         api.addDataAdapter("adapter://aleth.io/block/basic", new BlockBasicInfoAdapter(dataSource));
         api.addDataAdapter("adapter://aleth.io/full/block/details", new BlockDetailsAdapter(dataSource));
-        api.addDataAdapter("adapter://aleth.io/prices/latest", new NullEthPriceAdapter());
         api.addDataAdapter("adapter://aleth.io/block-range/summary", new BlockListAdapter(dataSource));
         api.addDataAdapter("adapter://aleth.io/block/confirmations",
             new BlockConfirmationsAdapter(dataSource.stores.blockStateStore));